+++ Modules/Intent.Modules.VisualStudio.Projects/Templates/LibraryCSProjectFile/Registrations.cs
@@ -1,7 +1,5 @@
-﻿using System.Collections.Generic;
-using System.ComponentModel;
+﻿using System.ComponentModel;
 using System.Linq;
-using Intent.Modules.Constants;
 using Intent.Engine;
 using Intent.Modules.VisualStudio.Projects.Api;
 using Intent.Registrations;
@@ -23,6 +21,10 @@ namespace Intent.Modules.VisualStudio.Projects.Templates.LibraryCSProjectFile
         public void DoRegistration(ITemplateInstanceRegistry registry, IApplication application)
         {
             var models = _metadataManager.VisualStudio(application).GetClassLibraryNETFrameworkModels();
+            if (models.Count > 0)
+            {
+                MsBuildLoaderHelper.EnsureMsBuildLoaded();
+            }
 
             foreach (var model in models)
             {